requirements.txt
# Binance API
ccxt==4.2.25
python-binance==1.0.19

# Data manipulation
pandas==2.1.4
numpy==1.26.3

# Technical analysis
ta==0.11.0

# Discord notifications
aiohttp==3.9.1

# Environment variables
python-dotenv==1.0.0

# Async support (included in Python 3.7+, but for clarity)
# asyncio - built-in

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.23.3

# Code quality (optional)
black==23.12.1
flake8==7.0.0
